refactor(test): organize import path
