+++ spec/scheduler/TestScheduler-spec.ts
@@ -4,8 +4,8 @@ import 'rxjs/add/operator/windowCount';
 import { ColdObservable } from 'rxjs/testing/ColdObservable';
 import { HotObservable } from 'rxjs/testing/HotObservable';
 import { SubscriptionLog } from 'rxjs/testing/SubscriptionLog';
-import { TestMessage } from '../../src/index';
 import { parseObservableMarble } from '../../src/marbles/parseObservableMarble';
+import { TestMessage } from '../../src/message/TestMessage';
 import { complete, error, next } from '../../src/message/TestMessageValue';
 import { TestScheduler } from '../../src/scheduler/TestScheduler';
